+++ CHANGELOG.md
@@ -1,10 +1,15 @@
+
+## 1.3.0
+Added `Marqueer.builder` for dynamic item build. [#1](https://github.com/GeceGibi/marqueer/issues/1)
+> When using `Marqueer.builder` if `itemCount` specified, `infinity` attribute must be `false` or if `itemCount` was a null `infinity` must be true.
+
 ## 1.2.1
 Fixed misspelled variable names. (thanks @Aboidrees)
 
 ## 1.2.0
 Improved desktop and web support.
 
-## 1.1.1
+p## 1.1.1
 Added `autoStartAfter` property to `Marquuer` widget for auto start after specific duration
 
 ## 1.1.0